Separated all six of the well.py classes into their own files. Reduced CC score to <=10 for all methods in all classes except BlockedWell.dataframe and BlockedWell.derive_from_dataframe. All original tests pass.
