+++ resqpy/well/__init__.py
@@ -1,11 +1,14 @@
-__all__ = ['md_datum', 'blocked_well', 'well_utils', 'add_wells_from_ascii_file', 'well_name', 'blocked_well_funcs', 'trajectory']
+__all__ = ['md_datum', 'blocked_well', 'add_wells_from_ascii_file', 'well_name', 'blocked_well_funcs', 'trajectory', 'deviation_survey', 'wellbore_frame', 'wellbore_frame_funcs']
 
 from .md_datum import MdDatum
 from .wellbore_marker_frame import WellboreMarkerFrame
-from .well_functions import DeviationSurvey, WellboreFrame
-from .well_utils import load_hdf5_array
+from .blocked_well import BlockedWell
+from .trajectory import Trajectory
+from .deviation_survey import DeviationSurvey
+from .wellbore_frame import WellboreFrame
+
 from .blocked_well_funcs import add_blocked_wells_from_wellspec, add_logs_from_cellio, lookup_from_cellio
 from .add_wells_from_ascii_file import add_wells_from_ascii_file
 from .well_name import well_name
-from .blocked_well import BlockedWell
-from .trajectory import Trajectory
+from .wellbore_frame_funcs import add_las_to_trajectory
+
