+++ tests/test_s_bend.py
@@ -145,7 +145,7 @@ def test_s_bend_fn(tmp_path, epc = None):
 
     df = df_trajectory(x, y, z)
 
-    datum = rqw.MdDatum(model, crs_uuid = crs.uuid, location = (x[0], y[0], z[0]))
+    datum = rqw2.MdDatum(model, crs_uuid = crs.uuid, location = (x[0], y[0], z[0]))
     datum.create_xml()
 
     trajectory = rqw2.Trajectory(model, md_datum = datum, data_frame = df, length_uom = 'm', well_name = 'ANGLED_WELL')
